DO.db {DO.db}	R Documentation

Bioconductor annotation data package

Description

Welcome to the DO.db annotation Package. The purpose of this package is to provide detailed information about the latest version of the Disease Ontology. This package is updated biannually.

You can learn what objects this package supports with the following command:

ls("package:DO.db")

Each of these objects has their own manual page detailing where relevant data was obtained along with some examples of how to use it.

Examples

  ls("package:DO.db")

[Package DO.db version 2.9 Index]