DOMAPCOUNTS {DO.db}	R Documentation

Number of mapped keys for the maps in package DO.db

Description

DOMAPCOUNTS provides the "map count" (i.e. the count of mapped keys) for each map in package DO.db.

Details

This "map count" information is precalculated and stored in the package annotation DB. This allows some quality control and is used by the checkMAPCOUNTS function defined in AnnotationDbi to compare and validate different methods (like count.mappedkeys(x) or sum(!is.na(as.list(x)))) for getting the "map count" of a given map.

See Also

mappedkeys, count.mappedkeys, checkMAPCOUNTS

Examples

  DOMAPCOUNTS
  mapnames <- names(DOMAPCOUNTS)
  DOMAPCOUNTS[mapnames[1]]
  x <- get(mapnames[1])
  sum(!is.na(as.list(x)))
  count.mappedkeys(x)   # much faster!

  ## Check the "map count" of all the maps in package DO.db
  #checkMAPCOUNTS("DO.db")

[Package DO.db version 2.9 Index]