DOOFFSPRING {DO.db}	R Documentation

Annotation of DO Identifiers to their Offspring

Description

This data set describes associations between DO terms and their offspring terms, based on the directed acyclic graph (DAG) defined by the Disease Ontology Consortium. The format is an R object mapping the DO terms to all offspring terms, where an ancestor term is a more specific DO term that is preceded by the given DO term in the DAG (in other words, the children and all their children, etc.).

Details

Each DO term is mapped to a vector of offspring DO terms.

Mappings were based on data provided by: Disease Ontology With a date stamp from the source of: 20150323 (sub_version 2806)

References

http://do-wiki.nubic.northwestern.edu/index.php/Main_Page

Examples

  # Convert the object to a list
  xx <- as.list(DOOFFSPRING)
  # Remove DO IDs that do not have any offspring
  xx <- xx[!is.na(xx)]
   if(length(xx) > 0){
    # Get the offspring DO identifiers for the first two elents of xx
    doidentifiers <- xx[1:2]
  }

[Package DO.db version 2.9 Index]