DOSYNONYM {DO.db}	R Documentation

Map from DO synonyms to DO terms

Description

DOSYNONYM is an R object that provides mapping from DO synonyms to DO terms

Details

Mappings were based on data provided by: Disease Ontology With a date stamp from the source of: 20150323 (sub_version 2806)

References

http://do-wiki.nubic.northwestern.edu/index.php/Main_Page

Examples

    x <- DOSYNONYM
    sample(x, 3)
    # DO ID "DOID:8757" has a lot of synonyms
    DOTERM[["DOID:8757"]]
    # DO ID "DOID:9134" is a synonym of DO ID "DOID:8757"
    DOID(DOSYNONYM[["DOID:9134"]])

[Package DO.db version 2.9 Index]