DOTerms-class {DO.db}	R Documentation

Class "DOTerms"

Description

A class to represent Disease Ontology nodes

Objects from the Class

Objects can be created by calls of the form DOTerms(DOId, term, synonym, secondary). DOId, term are required.

Slots

DOID:

Object of class "character" A character string for the DO id of a primary node.

Term:

Object of class "character" A character string that defines the role of gene product corresponding to the primary DO id.

Synonym:

Object of class "character" other ontology terms that are considered to be synonymous to the primary term attached to the DO id. Synonymous here can mean that the synonym is an exact synonym of the primary term, is related to the primary term, is broader than the primary term, is more precise than the primary term, or name is related to the term, but is not exact, broader or narrower.

Secondary:

Object of class "character" DO ids that are secondary to the primary DO id as results of merging DO terms so that One DO id becomes the primary DO id and the rest become the secondary.

Methods

DOID

signature(object = "DOTerms"): The get method for slot DOID.

Term

signature(object = "DOTerms"): The get method for slot Term.

Synonym

signature(object = "DOTerms"): The get method for slot Synonym.

Secondary

signature(object = "DOTerms"): The get method for slot Secondary.

show

signature(x = "DOTerms"): The method for pretty print.

Note

DOTerms objects are used to represent primary DO nodes in the SQLite-based annotation data package DO.db

References

http://do-wiki.nubic.northwestern.edu/index.php/Main_Page

Examples

  DOnode <- new("DOTerms", DOID="DOID:1234567", Term="Test")
  DOID(DOnode)
  Term(DOnode)

  ##Or you can just use these methods on a DOTermsAnnDbBimap
## Not run: ##I want to show an ex., but don't want to require DO.db
  require(DO.db)
  FirstTenDOBimap <- DOTERM[1:10] ##grab the 1st ten
  Term(FirstTenDOBimap)

  ##Or you can just use DO IDs directly
  ids = keys(FirstTenDOBimap)
  Term(ids)

## End(Not run)

[Package DO.db version 2.9 Index]