DO_dbconn {DO.db}	R Documentation

Collect information about the package annotation DB

Description

Some convenience functions for getting a connection object to (or collecting information about) the package annotation DB.

Usage

  DO_dbconn()
  DO_dbfile()

  DO_dbschema()
  DO_dbInfo()

Details

DO_dbconn returns a connection object to the package annotation DB. IMPORTANT: Don't call dbDisconnect on the connection object returned by DO_dbconn or you will break all the AnnDbObj objects defined in this package!

DO_dbfile returns the path (character string) to the package annotation DB (this is an SQLite file).

DO_dbschema prints the schema definition of the package annotation DB.

DO_dbInfo prints other information about the package annotation DB.

Value

DO_dbconn: a DBIConnection object representing an open connection to the package annotation DB.

DO_dbfile: a character string with the path to the package annotation DB.

DO_dbschema: none (invisible NULL).

DO_dbInfo: none (invisible NULL).

See Also

dbGetQuery, dbConnect, dbconn, dbfile, dbschema, dbInfo

Examples

  ## Count the number of rows in the "do_term" table:
  dbGetQuery(DO_dbconn(), "SELECT COUNT(*) FROM do_term")

  ## The connection object returned by DO_dbconn() was
  ## created with:
  dbConnect(SQLite(), dbname=DO_dbfile(), cache_size=64000,
  synchronous='off')

  #DO_dbschema()

#  DO_dbInfo()

[Package DO.db version 2.9 Index]